+++ lib/tls/openssl/server.c
@@ -340,7 +340,7 @@ lws_tls_server_new_nonblocking(struct lws *wsi, lws_sockfd_type accept_fd)
 	}
 
 	SSL_set_ex_data(wsi->ssl, openssl_websocket_private_data_index, wsi);
-	SSL_set_fd(wsi->ssl, accept_fd);
+	SSL_set_fd(wsi->ssl, (int)(long long)accept_fd);
 
 #ifdef USE_WOLFSSL
 #ifdef USE_OLD_CYASSL
